delete some classes ans retrieve wasadigi
