+++ src/main/java/ch/heigvd/res/chill/domain/wasadigi/Boxer.java
@@ -1,13 +1,13 @@
-package ch.heigvd.res.chill.domain.mhganh;
+package ch.heigvd.res.chill.domain.wasadigi;
 
 import ch.heigvd.res.chill.domain.IProduct;
 
 import java.math.BigDecimal;
 
-public class PunkIPA implements IProduct {
+public class Boxer implements IProduct {
 
-  public final static String NAME = "Punk IPA";
-  public final static BigDecimal PRICE = new BigDecimal(4.2);
+  public final static String NAME = "Boxer";
+  public final static BigDecimal PRICE = new BigDecimal(2.0);
 
   @Override
   public String getName() {
