+++ src/main/java/ch/heigvd/res/chill/domain/wasadigi/LaFranche.java
@@ -1,13 +1,13 @@
-package ch.heigvd.res.chill.domain.mhganh;
+package ch.heigvd.res.chill.domain.wasadigi;
 
 import ch.heigvd.res.chill.domain.IProduct;
 
 import java.math.BigDecimal;
 
-public class Boxer implements IProduct {
+public class LaFranche implements IProduct {
 
-  public final static String NAME = "Boxer";
-  public final static BigDecimal PRICE = new BigDecimal(2.0);
+  public final static String NAME = "LaFranche";
+  public final static BigDecimal PRICE = new BigDecimal(4.8);
 
   @Override
   public String getName() {
