+++ src/main/java/ch/heigvd/res/chill/domain/wasadigi/PunkIPA.java
@@ -1,13 +1,13 @@
-package ch.heigvd.res.chill.domain.mhganh;
+package ch.heigvd.res.chill.domain.wasadigi;
 
 import ch.heigvd.res.chill.domain.IProduct;
 
 import java.math.BigDecimal;
 
-public class LaCorne implements IProduct {
+public class PunkIPA implements IProduct {
 
-  public final static String NAME = "La Corne";
-  public final static BigDecimal PRICE = new BigDecimal(4.85);
+  public final static String NAME = "Punk IPA";
+  public final static BigDecimal PRICE = new BigDecimal(4.2);
 
   @Override
   public String getName() {
